% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fucntions.R
\name{dpath}
\alias{dpath}
\title{dpath: dir path setting for normal and test code}
\usage{
dpath(mpath)
}
\arguments{
\item{path}{dir path}
}
\value{
set the gdpath global variable
}
\description{
dpath: dir path setting for normal and test code
}
\examples{
\dontrun{
dpath("inst/extdata/")
}

}
