% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fucntions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read: Convert the csv file content into data frame.}
\format{A data frame with columns:
\describe{
\item{STATE}{State (as number)}
\item{ST_CASE}{State Case Number}
\item{VE_TOTAL}{Number of Vehicle Forms Submitted- ALL}
\item{VE_FORMS}{Number of Motor Vehicles in Transport (MVIT)}
\item{PVH_INVL}{Number of Parked/Working Vehicles}
\item{PEDS}{Number of Forms Submitted for Persons Not in Motor Vehicles}
\item{PERNOTMVIT}{Number of Persons Not in Motor Vehicles in Transport (MVIT) }
\item{PERMVIT}{Number of Persons in Motor Vehicles in Transport (MVIT) }
\item{PERSONS}{Number of Forms Submitted for Persons in Motor Vehicles}
\item{COUNTY}{County}
\item{CITY}{City}
\item{DAY}{Day of Crash}
\item{MONTH}{Month of Crash}
\item{YEAR}{Year of Crash}
\item{DAY_WEEK}{Day of Week}
\item{HOUR}{Hour of Crash}
\item{MINUTE}{Minute of Crash}
\item{NHS}{National Highway System}
\item{ROAD_FNC}{Roadway Function Class (discontinued)}
\item{ROUTE}{Route Signing}
\item{TWAY_ID}{Trafficway Identifier}
\item{TWAY_ID2}{Trafficway Identifier}
\item{MILEPT}{Milepoint}
\item{LATITUDE}{Latitude}
\item{LONGITUD}{Longitude}
\item{SP_JUR}{Special Jurisdiction}
\item{HARM_EV}{First Harmful Event}
\item{MAN_COLL}{Manner of Collision}
\item{RELJCT1}{Relation to Junction- Within Interchange Area}
\item{RELJCT2}{Relation to Junction- Specific Location }
\item{TYP_INT}{Type of Intersection}
\item{WRK_ZONE}{Work Zone}
\item{REL_ROAD}{Relation to Trafficway}
\item{LGT_COND}{Light Condition}
\item{WEATHER1}{Atmospheric Conditions}
\item{WEATHER2}{Atmospheric Conditions}
\item{WEATHER}{Atmospheric Conditions}
\item{SCH_BUS}{School Bus Related}
\item{RAIL}{Rail Grade Crossing Identifier}
\item{NOT_HOUR}{Hour of Notification}
\item{NOT_MIN}{Minute of Notification}
\item{ARR_HOUR}{Hour of Arrival at Scene}
\item{ARR_MIN}{Minute of Arrival at Scene}
\item{HOSP_HR}{Hour of EMS(Emergency Medical Service. e.g.Ambulance) Arrival at Hospital}
\item{HOSP_MN}{Minute of EMS(Emergency Medical Service. e.g.Ambulance) Arrival at Hospital}
\item{CF1}{Crash Level Factor 1}
\item{CF2}{Crash Level Factor 2}
\item{CF3}{Crash Level Factor 3}
\item{FATALS}{Fatalities}
\item{DRUNK_DR}{The DRUNK_DR data element on the Crash level was incorrectly derived on all person
types from 1999 through 2007. Since then, it was derived based on all person types rather than
based on Drivers only. Furthermore, the data element name (DRUNK_DR) implies that the
individual was drunk, however, it actually captures those individuals whom the police reported
alcohol involvement OR who tested positive for alcohol (i.e. their blood alcohol concentration
was .01 g/dL or greater). Beginning with the 2008 Final FARS data file, DRUNK_DR has been
derived for Drivers only}
}

Error: throws error if the supplied file "foo" is not found, then throws the following error message
"file foo does not exist"}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}

Column level understanding following link/pdf is referred.
\url{http://www.nber.org/fars/ftp.nhtsa.dot.gov/fars/FARS-DOC/Analytical User Guide/USERGUIDE-2015.pdf}
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{( sample : accident_2013.csv.bz2 )}
}
\value{
data frame with the columns, mentioned below.
}
\description{
This is utility function.
This is called internally, called from "fars_read_years"
not intended for user call.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}


}
