% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fucntions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename: For the given "year", format and return filename string(as it is stored in file system).}
\usage{
make_filename(year)
}
\arguments{
\item{year}{. sample : 2013}
}
\value{
file name : sample : accident_2013.csv.bz2
}
\description{
This is utility function.
This is called internally, by "fars_read_years", and "fars_map_state" function.
not intended for user call.
}
\examples{
\dontrun{
make_filename("2013")
}

}
